requirements.txt
torch
torchvision
pytorch_pretrained_vit
streamlit
pillow